pyproject.toml
[tool.poetry]
name = "elevenlabs-haystack"
version = "0.1.0"
description = "ElevenLabs Text-to-Speech components for Haystack."
authors = ["Andy <andychert@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
elevenlabs-s3 = "^0.1.0"
haystack-ai = "^2.6.0"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
